worker/requirements.txt
injector==0.22.0

pydantic==2.10.6
pydantic-settings==2.7.1

celery==5.4.0
redis==5.2.1

torch==2.2.2
openai-whisper==20240930
moviepy==2.1.2
numpy==1.26.4